Splitted Cell component into different components
