+++ admin-data-explorer-slinky/src/main/scala/net/wiringbits/webapp/utils/ui/web/components/pages/ExperimentalTablesPage.scala
@@ -11,5 +11,4 @@ import slinky.core.annotations.react
   val component: FunctionalComponent[Props] = FunctionalComponent[Props] { props =>
     ExperimentalTableWidget(props.api)
   }
-
 }
