+++ admin-data-explorer-slinky/src/main/scala/net/wiringbits/webapp/utils/ui/web/components/widgets/ExperimentalTable.scala
@@ -23,7 +23,7 @@ import slinky.core.facade.Fragment
   case class Props(response: AdminGetTableMetadataResponse)
 
   private lazy val useStyles: StylesHook[Styles[Theme, Unit, String]] = {
-    val stylesCallback: StyleRulesCallback[Theme, Unit, String] = theme =>
+    val stylesCallback: StyleRulesCallback[Theme, Unit, String] = _ =>
       StringDictionary(
         "table" -> CSSProperties()
           .setTableLayout(TableLayoutProperty.fixed)
@@ -34,21 +34,10 @@ import slinky.core.facade.Fragment
   val component: FunctionalComponent[Props] = FunctionalComponent[Props] { props =>
     val classes = useStyles(())
 
-    val columns = props.response.fields.map { field =>
-      Cell(field.name, props.response.name, isField = true)
-    }
+    val title = Subtitle(snakeCaseToUpper(props.response.name))
 
-    val rows = props.response.rows.map { row =>
-      mui
-        .TableRow(
-          row.data.map { cell =>
-            // Here I'm supossing that the first column belongs to table ID
-            if (row.data.indexOf(cell) == 0) Cell(cell.value, props.response.name, isNav = true)
-            else Cell(cell.value, props.response.name)
-          }
-        )
-    }
-
+    val columns = props.response.fields.map(field => TableField(field.name))
+    val rows = props.response.rows.map(row => TableRow(row, props.response.name))
     val table =
       mui
         .Table(
@@ -64,7 +53,7 @@ import slinky.core.facade.Fragment
     Container(
       maxWidth = Some("100%"),
       child = Fragment(
-        Subtitle(snakeCaseToUpper(props.response.name)),
+        title,
         table,
         Pagination(props.response)
       )
