+++ admin-data-explorer-slinky/src/main/scala/net/wiringbits/webapp/utils/ui/web/components/widgets/PaginationWidget.scala
@@ -30,7 +30,7 @@ import scala.scalajs.js.|
         page = props.currentPage,
         rowsPerPage = props.pageLimit
       )
-      .onChangeRowsPerPage(event => props.onChangeRowsPerPage(event))
+      .onChangeRowsPerPage(props.onChangeRowsPerPage(_))
       .rowsPerPageOptions(pageOptions)
   }
 }
